refactor: split UI into components
